cleared warnings

+++ Flamingo/Flamingo/Source/animation.cpp
@@ -14,9 +14,9 @@ animation::animation(al::sprite *Sprite, int Frames, int FramesizeX, int Framesi
 	m_timer = 0.0f;
 	setVisibleFrame();
 #if _DEBUG
-	if (fmod(Sprite->getTextureSize().x,FramesizeX) != 0)
+	if (fmod(Sprite->getTextureSize().x,float(FramesizeX)) != 0)
 		std::cout<<"Check FramesizeX!!!!!!"<<std::endl;
-	if (fmod(Sprite->getTextureSize().y,FramesizeY) != 0)
+	if (fmod(Sprite->getTextureSize().y,float(FramesizeY)) != 0)
 		std::cout<<"Check FramesizeY!!!!!!"<<std::endl;
 #endif
 }
@@ -44,7 +44,7 @@ void animation::update(float deltaTime)
 			else
 				m_currentFrame--;
 		}
-		m_timer -= 1.0 / m_fps;
+		m_timer -= 1.0f / m_fps;
 		//counter++;
 		
 		setVisibleFrame();
@@ -75,7 +75,7 @@ void animation::setVisibleFrame()
 	
 	int currentFrameX = m_currentFrame % ((int)((m_sprite->getTextureSize().x / m_framesizeX)+0.5f)) * m_framesizeX;
 	int currentFrameY = (int)(floor((double)m_currentFrame / ((double)m_sprite->getTextureSize().x / (double)m_framesizeX))) * m_framesizeY;
-	al::rectangle frame = al::rectangle(currentFrameX, currentFrameY, m_framesizeX, m_framesizeY);
+	al::rectangle frame = al::rectangle((float)currentFrameX, (float)currentFrameY, (float)m_framesizeX, (float)m_framesizeY);
 	m_sprite->setTextureRectangle(frame);
 }
